How to use: * Use the `StunStars` prefab * Set `FollowTarget` and `FollowOffset` properties - there's no lerp, the stars will follow the target instantly * Use the `void Play(float duration)` function to trigger. Note that `duration` can't be shorter than `AnimTime` (whose default is 2 seconds).
When duration is longer than AnimTime, it plays the curve animations for half of AnimTime, then pauses the curves for (duration - AnimTime) seconds, then plays the last half of the animations. The curves only control rates of change, so the animation doesn't really stop it just gets extended.

I can show Cody how to control the curves for the animation if she wants to tweak them. Someone can also make a better model.