Things are all pretty much at a constant distance from the camera
So I removed shadow cascades so that we have consistent shadow quality across the entire scene (https://docs.unity3d.com/Manual/DirLightShadows.html)

It also helps for the eventual low LOD camera. That camera needs to be further from the scene to negate its zoom - and that caused havoc with shadow cascades because Unity was selecting low quality shadows because of the distance.