Someone added this task to my trello: Improve freeze visuals if possible (some clipping currently, grass is still green)
I improved visuals as much as I could. It's better now but some jankiness remains because of object scale animation not being compatible with TAA out of the box. Too long to fix, we can live with it.

Some screenshots of the clipping would help because it looks alright now. Maybe the new arena layout fixed it?

Finally i made the grass turn cyan when frozen