While items are carried atm their movement is jerky.
I tried finding the root cause but couldn't. In the end what works well is to set the ingredient to the parent position directly with no lerp.

So now once an item has been picked up the lerp gradually disappears until finally it reaches "direct set to parent position" mode.