When huge forces are involved then things can manage to fly through our bound colliders.
One way to solve it is to set continuous collision detection on our RigidBodies.

Another way is to add this script to our ingredient and player prefabs (script should be on the same object as the RigidBody component). I haven't added it to the prefabs to avoid merge conflicts, it can be done later. I did test it though, works great.